Title: 10개 입력받기, Time: 123ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240715/10개 입력받기/receive-10-inputs.java
@@ -4,7 +4,7 @@ public class Main {
         // 여기에 코드를 작성해주세요.
         Scanner sc = new Scanner(System.in);
         int [] score = new int [10];
-        int cnt = 0;
+        double cnt = 0;
         int sum = 0;
         double avg = 0.0;
